+++ deployment/deploy-s3-dist.sh
@@ -14,6 +14,8 @@ DEPLOY_DIR="$PWD"
 SOURCE_DIR="$DEPLOY_DIR/../source"
 TEMPLATE_DIST_DIR="global-s3-assets"
 BUID_DIST_DIR="regional-s3-assets"
+MAIN_TEMPLATE="amazon-rekognition-custom-brand-detection.template"
+
 
 #
 # @function usage
@@ -27,7 +29,7 @@ It should be run from the repo's deployment directory
 
 ------------------------------------------------------------------------------
 cd deployment
-bash ./deploy-s3-dist.sh --bucket BUCKET_NAME [--acl ACL_SETTING] [--profile AWS_PROFILE] [--region AWS_REGION]
+bash ./deploy-s3-dist.sh --bucket BUCKET_NAME [--acl ACL_SETTING] [--profile AWS_PROFILE]
 
 where
   --bucket BUCKET_NAME        specify the bucket name where the templates and packages deployed to.
@@ -38,8 +40,6 @@ where
                               'public-read' acl settings
 
   --profile AWS_PROFILE       [optional] specify the AWS CLI profile. If not specified, it assumes 'default'
-
-  --region AWS_REGION         [optional] specify AWS_REGION. If not specified, it assumes 'us-east-1'
 "
   return 0
 }
@@ -80,11 +80,6 @@ while [[ $# -gt 0 ]]; do
       shift # past argument
       shift # past value
       ;;
-      -r|--region)
-      REGION="$2"
-      shift # past argument
-      shift # past value
-      ;;
       *)
       shift
       ;;
@@ -116,9 +111,6 @@ done
 [ -z "$PROFILE" ] && \
   PROFILE="default"
 
-[ -z "$REGION" ] && \
-  REGION="us-east-1"
-
 ACCOUNTID=$(aws sts get-caller-identity | jq .Account | tr -d \")
 [ -z "$ACCOUNTID" ] && \
   echo "error: fail to get AWS Account ID" && \
@@ -142,20 +134,44 @@ function copy_to_bucket() {
   [ -z "$location" ] && \
     echo "Bucket '${bucket}' either doesn't exist or doesn't belong to accountId '${ACCOUNTID}'. exiting..." && \
     exit 1
+
   local region="us-east-1"
   [ "$location" != "null" ] && \
     region=$location
 
+  local domain="s3.amazonaws.com"
+  local optionalFlag="--acl ${ACL_SETTING} --profile ${PROFILE}"
+
+  if [ "$region" != "us-east-1" ]; then
+    domain=s3.${region}.amazonaws.com
+    optionalFlag="${optionalFlag} --region ${region}"
+  fi
+
   # upload artifacts to bucket
   echo "== Deploy '${SOLUTION} ($VERSION)' package from '${fullPackages}' to '${versionFolder}' in '${region}' [BEGIN] =="
-  if [ "$region" == "us-east-1" ]; then
-    aws s3 cp $fullPackages $versionFolder --recursive --acl ${ACL_SETTING} --profile ${PROFILE}
-    aws s3 cp $mainTemplate $latestFolder --recursive --acl ${ACL_SETTING} --profile ${PROFILE}
-  else
-    aws s3 cp $fullPackages $versionFolder --recursive --acl ${ACL_SETTING} --profile ${PROFILE} --region ${region}
-    aws s3 cp $mainTemplate $latestFolder --recursive --acl ${ACL_SETTING} --profile ${PROFILE} --region ${region}
-  fi
+  aws s3 cp $fullPackages $versionFolder --recursive $optionalFlag
+  aws s3 cp $mainTemplate $latestFolder --recursive $optionalFlag
   echo "== Deploy '${SOLUTION} ($VERSION)' package from '${fullPackages}' to '${versionFolder}' in '${region}' [COMPLETED] =="
+
+  local url="https://${bucket}.${domain}/${SOLUTION}/${VERSION}/${MAIN_TEMPLATE}"
+  local latestUrl="https://${bucket}.${domain}/${SOLUTION}/latest/${MAIN_TEMPLATE}"
+
+  echo "== (VERSIONED URL) ============================"
+  echo ""
+  echo "HTTPS URL:"
+  echo "$url"
+  echo ""
+  echo "One-click URL to create stack:"
+  echo "https://console.aws.amazon.com/cloudformation/home?region=${region}#/stacks/quickcreate?templateURL=${url}&stackName=custom-brand"
+  echo ""
+
+  echo "== (LATEST URL) ==============================="
+  echo ""
+  echo "$latestUrl"
+  echo ""
+  echo "One-click URL to create stack:"
+  echo "https://console.aws.amazon.com/cloudformation/home?region=${region}#/stacks/quickcreate?templateURL=${latestUrl}&stackName=custom-brand"
+  echo ""
 }
 
 #