+++ source/analysis-video/states/create-sprite-images/index.js
@@ -1,12 +1,15 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 const PATH = require('path');
+const Jimp = require('jimp');
 const {
   States,
   S3Utils,
 } = require('core-lib');
 const SpriteHelper = require('../shared/spriteHelper');
 const BaseState = require('../shared/baseState');
+
+const BORDER_SIZE = 1;
 
 class StateCreateSpriteImages extends BaseState {
   async process() {
@@ -43,11 +46,12 @@ class StateCreateSpriteImages extends BaseState {
       maxPerRow,
     ] = SpriteHelper.computeSpriteSize(data.streams[0].width, data.streams[0].height);
 
-    const canvasLib = require('canvas');
-    const [
-      canvas,
-      context,
-    ] = this.createSpriteCanvas(canvasLib, data.frames.length, spriteW, spriteH, maxPerRow);
+    const combined = await this.createSpriteBackground(
+      data.frames.length,
+      spriteW,
+      spriteH,
+      maxPerRow
+    );
 
     let rowIdx = 0;
     while (data.frames.length) {
@@ -55,10 +59,19 @@ class StateCreateSpriteImages extends BaseState {
       for (let colIdx = 0; colIdx < splices.length; colIdx++) {
         const frame = splices[colIdx];
         const key = PATH.join(prefix, `${frame.coded_picture_number}.jpg`);
-        const img = await this.readImage(canvasLib, src.bucket, key);
-        context.drawImage(img,
-          0, 0, img.width, img.height,
-          (colIdx * spriteW), (rowIdx * spriteH), spriteW, spriteH);
+
+        // scale and crop image
+        const img = await this.readImage(src.bucket, key);
+        const factor = spriteW / img.bitmap.width;
+        const scaled = img.scale(factor);
+
+        const w = scaled.bitmap.width - (BORDER_SIZE * 2);
+        const h = scaled.bitmap.height - (BORDER_SIZE * 2);
+        const cropped = scaled.crop(BORDER_SIZE, BORDER_SIZE, w, h);
+
+        const l = colIdx * spriteW + BORDER_SIZE;
+        const t = rowIdx * spriteH + BORDER_SIZE;
+        combined.blit(cropped, l, t);
       }
       rowIdx++;
     }
@@ -66,10 +79,10 @@ class StateCreateSpriteImages extends BaseState {
     prefix = this.makeOutputPath(src.key, States.CreateSpriteImages);
     const name = `${src.index}.jpg`;
     const key = PATH.join(prefix, name);
-    const image = canvas.toBuffer('image/jpeg', {
-      quality: 0.95,
-    });
-    await S3Utils.upload(src.bucket, key, image, {
+
+    const jpeg = await combined.quality(80).getBufferAsync(Jimp.MIME_JPEG);
+
+    await S3Utils.upload(src.bucket, key, jpeg, {
       ContentType: 'image/jpeg',
       ContentDisposition: `attachment; filename="${name}"`,
     });
@@ -84,31 +97,46 @@ class StateCreateSpriteImages extends BaseState {
     return result;
   }
 
-  createSpriteCanvas(canvasLib, totalFrames, spriteW, spriteH, maxPerRow) {
-    const cols = Math.min(totalFrames, maxPerRow);
-    let rows = Math.floor(totalFrames / maxPerRow);
-    if ((totalFrames % maxPerRow) > 0) {
-      rows++;
-    }
-    console.log(`canvas: ${cols}x${rows} [${spriteW * cols}x${spriteH * rows}] (${totalFrames})`);
-    const canvas = canvasLib.createCanvas(spriteW * cols, spriteH * rows);
-    return [
-      canvas,
-      canvas.getContext('2d'),
-    ];
+  async createSpriteBackground(
+    totalFrames,
+    spriteW,
+    spriteH,
+    maxPerRow
+  ) {
+    return new Promise((resolve, reject) => {
+      const cols = Math.min(totalFrames, maxPerRow);
+      let rows = Math.floor(totalFrames / maxPerRow);
+      if ((totalFrames % maxPerRow) > 0) {
+        rows++;
+      }
+
+      const imgW = spriteW * cols;
+      const imgH = spriteH * rows;
+
+      console.log(`background: ${cols}x${rows} [${imgW}x${imgH}] (${totalFrames})`);
+
+      // eslint-disable-next-line
+      const _ = new Jimp(imgW, imgH, 0x000000ff, (e, image) => {
+        if (e) {
+          reject(e);
+        } else {
+          resolve(image);
+        }
+      });
+    });
   }
 
-  async readImage(canvasLib, bucket, key) {
-    const buf = await S3Utils.getObject(bucket, key)
-      .then(data => data.Body);
+  async readImage(bucket, key) {
     return new Promise((resolve, reject) => {
-      const img = new canvasLib.Image();
-      img.onload = () => resolve(img);
-      img.onerror = (e) => {
-        console.log(`readImage(${key}): ${e.message}`);
-        reject(e);
-      };
-      img.src = buf;
+      const signed = S3Utils.signUrl(bucket, key);
+      Jimp.read(signed)
+        .then((img) => {
+          resolve(img);
+        })
+        .catch((e) => {
+          console.log(`ERR: readImage: ${key}: ${e.message}`);
+          reject(e);
+        });
     });
   }
 }