+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-0423
+### Changed
+- update to NodeJS 20.x
+- update ffmpeg dependencies
+- replace Canvas with Jimp library
+- fix bucket AccessControl
+
+### Removed
+- Remove canvas library
+
+### Added
+- Added Jimp library for image processing
+
 ## [1.3.1] - 2021-1021
 ### Changed
 - remove popper dependency